+++ src/main/java/tree/Addition.java
@@ -1,3 +1,7 @@
+package tree;
+
+import visitor.Visitor;
+
 public class Addition extends OperateurBinaire {
     public Addition(Noeud ng, Noeud nd) {
         super("+", ng, nd);
