+++ src/main/java/tree/Multiplication.java
@@ -1,3 +1,7 @@
+package tree;
+
+import visitor.Visitor;
+
 public class Multiplication extends OperateurBinaire {
 
     public Multiplication(Noeud ng, Noeud nd) {
