+++ src/main/java/tree/Negation.java
@@ -1,3 +1,7 @@
+package tree;
+
+import visitor.Visitor;
+
 public class Negation extends OperateurUnaire {
     public Negation(Noeud n) {
         super("-", n);
