+++ src/main/java/tree/Noeud.java
@@ -1,3 +1,7 @@
+package tree;
+
+import visitor.Visitor;
+
 public interface Noeud
 {
     void accept(Visitor visitor);
