+++ src/main/java/tree/OperateurUnaire.java
@@ -1,3 +1,5 @@
+package tree;
+
 public abstract class OperateurUnaire implements Noeud {
     private String op;
     private Noeud opG;
