+++ src/main/java/visitor/InfixeRoute.java
@@ -1,7 +1,14 @@
+package visitor;
+
+import tree.Constante;
+import tree.Negation;
+import tree.OperateurUnaire;
+import tree.OperateurBinaire;
+
 /**
  * Created by yannick on 22/03/17.
  */
-public class PostfixRoute implements Visitor {
+public class InfixeRoute implements Visitor {
 
 
     public void visitOperateurUnaire(OperateurUnaire opUnaire)
@@ -13,8 +20,8 @@ public class PostfixRoute implements Visitor {
     public void visitOperateurBinaire(OperateurBinaire opBinaire)
     {
         opBinaire.getOpG().accept(this);
-        opBinaire.getOpD().accept(this);
         System.out.println(opBinaire.getOp());
+        opBinaire.getOpD().accept(this);
     }
 
     public void visitConstante(Constante constante)
@@ -24,7 +31,7 @@ public class PostfixRoute implements Visitor {
 
     public void visitNegation(Negation neg)
     {
+        System.out.println(String.format("(%s)",neg.getOp()));
         neg.getOpG().accept(this);
-        System.out.println(neg.getOp());
     }
 }
