+++ src/main/java/visitor/VisitorCalculateExpression.java
@@ -1,7 +1,14 @@
+package visitor;
+
+import tree.Constante;
+import tree.Negation;
+import tree.OperateurUnaire;
+import tree.OperateurBinaire;
+
 /**
  * Created by yannick on 22/03/17.
  */
-public class PrefixRoute implements Visitor {
+public class VisitorCalculateExpression implements Visitor {
 
 
     public void visitOperateurUnaire(OperateurUnaire opUnaire)
